requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
scikit-learn==1.3.0
joblib==1.3.0
numpy==1.24.0
pandas==2.0.0
python-dotenv==1.0.0
nltk==3.8
textblob==0.17.1
google-generativeai==0.8.0
scipy==1.11.0
